backend/requirements.txt
fastapi==0.111.0
mangum==0.17.0
uvicorn==0.29.0
pydantic==2.7.1
boto3==1.34.100
joblib==1.4.2
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
nltk==3.8.1
lime==0.2.0.1
pandas
# Local development on macOS arm64 (Apple Silicon).
# Docker/Lambda uses requirements-docker.txt with tensorflow-cpu instead.
tensorflow-macos==2.16.2
